refactor: 유저의 입력을 받는 함수 분리
