+++ src/index.js
@@ -1,35 +1,7 @@
-import { readLineAsync } from "./input.js";
 import { createGameState } from "./store/createGameState.js";
-import { validateNumber, validateRange } from "./utils/index.js";
+import { validateNumber } from "./utils/index.js";
 import { displayGameResult, displayHint } from "./display/index.js";
-
-async function initializeRange() {
-  while (true) {
-    try {
-      console.log("[게임 설정] 게임 시작을 위해 최소 값, 최대 값을 입력해주세요. (예: 1, 50)");
-      const rangeInput = await readLineAsync("숫자 입력: ");
-      const { min, max } = validateRange(rangeInput);
-      return { min, max };
-    } catch (error) {
-      console.log(error.message);
-      console.log("다시 입력해주세요.");
-    }
-  }
-}
-
-async function initializeMaxAttempts() {
-  while (true) {
-    try {
-      console.log("[게임 설정] 게임 시작을 위해 진행 가능 횟수를 입력해주세요.");
-      const attemptsInput = await readLineAsync("숫자 입력: ");
-      const maxAttempts = validateNumber(attemptsInput);
-      return maxAttempts;
-    } catch (error) {
-      console.log(error.message);
-      console.log("다시 입력해주세요.");
-    }
-  }
-}
+import { readLineAsync, askRange, askMaxAttempts, askToPlayAgain } from "./user/index.js";
 
 async function play(gameState) {
   console.log(
@@ -66,25 +38,9 @@ async function play(gameState) {
   }
 }
 
-async function askToPlayAgain() {
-  while (true) {
-    const answer = await readLineAsync("게임을 다시 시작하시겠습니까? (yes/no): ");
-    const lowerAnswer = answer.toLowerCase();
-
-    if (lowerAnswer === "yes") {
-      return true;
-    }
-    if (lowerAnswer === "no") {
-      return false;
-    }
-
-    console.log("yes 또는 no만 입력해주세요.");
-  }
-}
-
 async function upAndDownGame() {
-  const { min, max } = await initializeRange();
-  const maxAttempts = await initializeMaxAttempts();
+  const { min, max } = await askRange();
+  const maxAttempts = await askMaxAttempts();
 
   const gameState = createGameState(min, max, maxAttempts);
