+++ src/user/askToPlayAgain.js
@@ -1,0 +1,17 @@
+async function askToPlayAgain() {
+  while (true) {
+    const answer = await readLineAsync("게임을 다시 시작하시겠습니까? (yes/no): ");
+    const lowerAnswer = answer.toLowerCase();
+
+    if (lowerAnswer === "yes") {
+      return true;
+    }
+    if (lowerAnswer === "no") {
+      return false;
+    }
+
+    console.log("yes 또는 no만 입력해주세요.");
+  }
+}
+
+export default askToPlayAgain
